feat: create separate pipelines module
